move LoginUser method to UsersHandlers
